Location is now updated automatically when the http succeeeds.
