+++ dvd_categorizer_web/Cargo.toml
@@ -12,6 +12,7 @@ dioxus = { version = "0.7.3", features = ["router"] }
 log.workspace = true
 env_logger.workspace = true
 tracing.workspace = true
+wasm-logger = "0.2"
 dotenvy.workspace = true
 models = { path = "../models", features = [] }
 reqwest = { version = "0.12.15", default-features = false, features = ["json"] }
