+++ dvd_categorizer_web/src/components/mod.rs
@@ -1,3 +1,4 @@
 pub mod display_movie;
+pub mod location_editor;
 pub mod movie_card;
 pub mod movie_grid;
